💥 Drop deprecated `.noShrink`
