+++ packages/fast-check/src/arbitrary/noShrink.ts
@@ -1,4 +1,23 @@
-import type { Arbitrary } from '../check/arbitrary/definition/Arbitrary';
+import type { Value } from '../check/arbitrary/definition/Value';
+import type { Random } from '../random/generator/Random';
+import { Arbitrary } from '../check/arbitrary/definition/Arbitrary';
+import { Stream } from '../stream/Stream';
+
+/** @internal */
+class NoShrinkArbitrary<T> extends Arbitrary<T> {
+  constructor(readonly arb: Arbitrary<T>) {
+    super();
+  }
+  generate(mrng: Random, biasFactor: number | undefined): Value<T> {
+    return this.arb.generate(mrng, biasFactor);
+  }
+  canShrinkWithoutContext(value: unknown): value is T {
+    return this.arb.canShrinkWithoutContext(value);
+  }
+  shrink(_value: T, _context?: unknown): Stream<Value<T>> {
+    return Stream.nil();
+  }
+}
 
 /**
  * Build an arbitrary without shrinking capabilities.
@@ -14,5 +33,5 @@ import type { Arbitrary } from '../check/arbitrary/definition/Arbitrary';
  * @public
  */
 export function noShrink<T>(arb: Arbitrary<T>): Arbitrary<T> {
-  return arb.noShrink();
+  return new NoShrinkArbitrary(arb);
 }
