+++ packages/fast-check/test/unit/check/arbitrary/definition/Arbitrary.utest.spec.ts
@@ -663,75 +663,4 @@ describe('NextArbitrary', () => {
       expect(shrink).not.toHaveBeenCalled();
     });
   });
-
-  describe('noShrink', () => {
-    it('should simply return the original instance of Value on generate', () => {
-      // Arrange
-      const expectedBiasFactor = 48;
-      const generate = vi.fn();
-      const canShrinkWithoutContext = vi.fn() as any as (value: unknown) => value is any;
-      const shrink = vi.fn();
-      const choice = new Value(1, Symbol());
-      generate.mockReturnValueOnce(choice);
-      class MyNextArbitrary extends Arbitrary<any> {
-        generate = generate;
-        canShrinkWithoutContext = canShrinkWithoutContext;
-        shrink = shrink;
-      }
-
-      // Act
-      const arb = new MyNextArbitrary().noShrink();
-      const g = arb.generate(mrngNoCall, expectedBiasFactor);
-
-      // Assert
-      expect(g).toBe(choice); // just returning the instance of the source arbitrary (including its context)
-      expect(generate).toHaveBeenCalledWith(mrngNoCall, expectedBiasFactor);
-    });
-
-    it('should override default shrink with function returning an empty Stream', () => {
-      // Arrange
-      const shrink = vi.fn();
-      class MyNextArbitrary extends Arbitrary<any> {
-        generate(): Value<any> {
-          throw new Error('Not implemented.');
-        }
-        canShrinkWithoutContext(value: unknown): value is any {
-          throw new Error('Not implemented.');
-        }
-        shrink = shrink;
-      }
-      const fakeArbitrary: Arbitrary<any> = new MyNextArbitrary();
-      const noShrinkArbitrary = fakeArbitrary.noShrink();
-
-      // Act
-      const out = noShrinkArbitrary.shrink(5, Symbol());
-
-      // Assert
-      expect([...out]).toHaveLength(0);
-      expect(shrink).not.toHaveBeenCalled();
-    });
-
-    it('should return itself when called twice', () => {
-      // Arrange
-      class MyNextArbitrary extends Arbitrary<any> {
-        generate(): Value<any> {
-          throw new Error('Not implemented.');
-        }
-        canShrinkWithoutContext(value: unknown): value is any {
-          throw new Error('Not implemented.');
-        }
-        shrink(): Stream<Value<any>> {
-          throw new Error('Not implemented.');
-        }
-      }
-      const fakeArbitrary: Arbitrary<any> = new MyNextArbitrary();
-
-      // Act
-      const firstNoShrink = fakeArbitrary.noShrink();
-      const secondNoShrink = firstNoShrink.noShrink();
-
-      // Assert
-      expect(secondNoShrink).toBe(firstNoShrink);
-    });
-  });
 });
